Move point/line interfaces; factor out rotate-point function
