+++ src/find-symmetry-lines.ts
@@ -1,3 +1,5 @@
+import { Point, Line } from './interfaces';
+
 // Toggle debug logging.
 const DEBUG = false;
 
@@ -328,17 +330,6 @@ function debug(fn: () => void) {
   if (DEBUG) {
     fn();
   }
-}
-
-export interface Point {
-  x: number;
-  y: number;
-}
-
-export interface Line {
-  // A line is defined by two points, which should not be the same.
-  p1: Point;
-  p2: Point;
 }
 
 // Used in tests. Consider refactoring.
